\name{ClearWorkspace}
\alias{ClearWorkspace}
\title{  ClearWorkspace() removes all Objects from Workspace usufull to delete older versions of functions etc }
\arguments{
}
\value{List with
}
\description{
  ClearWorkspace() removes all Objects from Workspace usufull to delete older versions of functions etc }
\author{
}
