% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_RRHO_file_intersect_then_rank.R
\name{create_RRHO_file_intersect_then_rank}
\alias{create_RRHO_file_intersect_then_rank}
\title{create RRHO file after intersecting}
\usage{
create_RRHO_file_intersect_then_rank(metric_file1, metric_file2,
  outfile.prefix = paste0("rankrank.", metric_file1, "_vs_",
  metric_file2), metric.col.file1 = 2, metric.col.file2 = 2,
  rank1 = "desc", rank2 = "desc")
}
\arguments{
\item{metric_file1}{filepath of first metric file. usually 2 columns. should have no headers. first column is gene name. second column is metric. does not need to be sorted}

\item{metric_file2}{filepath of second metric file. usually 2 columns. should have no headers. first column is gene name. second column is metric. does not need to be sorted}

\item{metric.col.file1}{# of the column that metric data is in for file 1}

\item{metric.col.file2}{# of the column that metric data is in for file 2}

\item{rank1}{default is descending rank (i.e. higher metric is rank 1)}

\item{rank2}{default is descending rank (i.e. higher metric is rank 1)}

\item{outfile_prefix}{prefix of the output file}
}
\description{
create RRHO file after intersecting
}
