\name{dbt.general-package}
\alias{dbt.general-package}
\alias{dbt.general}
\docType{package}
\title{
Basic function for data manipulation.
}
\description{
You can rearrange matrices/vectors, reorder values, calculate the logarithm of z-transformed values,
get the lower/upper triangular matrix as a vector etc.
}
\details{
\tabular{ll}{
Package: \tab dbt.general\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2009-05-12\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
Functions:

eqfloat()

maxmtx()

normstrich()

norm2strich()

norm3strich()

rsample()

sigmoid()

sortby()

sortdecending()

tail()

trilvec()

triuvec()

uniquePoints()

zinvlog()

zlog()
}
\author{
Sascha Feld

Maintainer: AG Datenbionik, ultsch@mathematik.uni-marburg.de
}
%\references{
% ~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file ~~
%basic function, matrix manipulation, matrix rearrangement, vector manipulation,
%vector rearrangement, z-transformed vlaues log, reorder values
%\keyword{ package }
%\seealso{
%----------------- HIER ggf. die Main-Page zum Gesamprojekt einfügen.
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
