% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matches.R
\name{find_packages}
\alias{find_packages}
\title{Search through files for packages used; make sure DESCRIPTION is up to date.}
\usage{
find_packages(pkg_path, output_path)
}
\description{
Search through files for packages used; make sure DESCRIPTION is up to date.
}
