\name{findAttrCol}
\alias{findAttrCol}
\title{ Find attribute column of data matrix by name }
\arguments{
\item{AttrName}{attribute name, trailing spaces can be omitted}
\item{VarNames}{array with column names}
\item{Data}{data matrix}
}
\value{AttrCol   attribute column vektor}
\description{
 Find attribute column of data matrix by name }
\author{
}
