\name{findAttrInd}
\alias{findAttrInd}
\title{ Find attribute column index of data matrix by name dbt/Attributes }
\arguments{
\item{AttrName}{attribute name, trailing spaces can be omitted}
\item{VarNames}{array with column names}
}
\value{AttrInd   attribute column index}
\description{
 Find attribute column index of data matrix by name dbt/Attributes }
\author{
}
