% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-package.R
\docType{package}
\name{general-package}
\alias{general}
\alias{general-package}
\title{general: An Package containing general functions 0804}
\description{
A simple general package to demonstrate package creation 0804.
}
\author{
\strong{Maintainer}: Eric Chen \email{cysh010479@gmail.com}

}
\keyword{internal}
