% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-flat-violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{ggplot Flat Violin}
\usage{
geom_flat_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", trim = TRUE, scale = "area",
  show.legend = NA, inherit.aes = TRUE, ...)
}
\description{
ggplot Flat Violin
}
\details{
Copy-pasted from https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d/raw/fb53bd97121f7f9ce947837ef1a4c65a73bffb3f/geom_flat_violin.R
somewhat hackish solution to:
https://twitter.com/EamonCaddigan/status/646759751242620928
based mostly on copy/pasting from ggplot2 geom_violin source:
https://github.com/hadley/ggplot2/blob/master/R/geom-violin.r
}
\examples{
:
ggplot(diamonds, aes(cut, carat)) +
  geom_flat_violin() +
  coord_flip()
}
