\name{getVarByHeader}
\alias{getVarByHeader}
\title{  getVarByHeader(V,i) Selectes Variable by Name in Header}
\arguments{
\item{V}{list of ReadLRN or ReadData}
\item{i}{which Variable do you want to select}
}
\value{Var     Vecotor of Variable wit Varname=Header[i] in Global Environment}
\description{
  getVarByHeader(V,i) Selectes Variable by Name in Header}
\author{
MT 08/2015}
