% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulate ggplot colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{the number of colors needed}
}
\description{
from https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
