\name{intersectIND}
\alias{intersectIND}
\title{ Schnitt der der beiden Vektoren sowie die zugehörigen Indizes um Datensätze vergleichen zu können V <- intersectKeys(vector1,vector2) Intersection= V$intersection Ind2First   = V$ind1   ; an index such that intersection == vector1[ind1] Ind2Second  = V$ind2   ; an index such that intersection == vector2[ind2] }
\arguments{
\item{vector1[1:n]}{numerischer Vektor}
\item{vector[1:m]}{numerischer Vektor}
\item{fast}{=TRUE Bei sehr häufiger verwendung, kann spezielles Paket geladen werden, welches eine schellere Ausführung ermöglichst}
}
\value{V$intersection[1:p]      numerischer Vektor des schnittes}
\description{
 Schnitt der der beiden Vektoren sowie die zugehörigen Indizes um Datensätze vergleichen zu können V <- intersectKeys(vector1,vector2) Intersection= V$intersection Ind2First   = V$ind1   ; an index such that intersection == vector1[ind1] Ind2Second  = V$ind2   ; an index such that intersection == vector2[ind2] }
\author{
:}
