\name{lustbutone}
\alias{lustbutone}
\title{ returns the last and last but one element of an vector author: MT exmaple           ende=lustbutone(Stressverlauf)           letztereintrag=ende$Last           vorletztereintragende$Lastbutone}
\arguments{
}
\value{List with
}
\description{
 returns the last and last but one element of an vector author: MT exmaple           ende=lustbutone(Stressverlauf)           letztereintrag=ende$Last           vorletztereintragende$Lastbutone}
\author{
}
