\name{requireRpackage}
\alias{requireRpackage}
\title{  requireRpackage(name) Loads or automatically installs and loads a R package. If package cant be installed, the function stops }
\arguments{
\item{name}{string with name of package}
\item{biocite}{=FALSE, if TRUE Loading Package from biocite instead of CRAN}
\item{Verbose}{=TRUE: all 'simple' diagnostic messages on package loading}
}
\value{List with
}
\description{
  requireRpackage(name) Loads or automatically installs and loads a R package. If package cant be installed, the function stops }
\author{
MT 09/2014}
