% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-create-doc.R
\name{rmd_create_doc}
\alias{rmd_create_doc}
\title{Render the Rmarkdown file and move the output}
\usage{
rmd_create_doc(run.dir, rmd.file, final.dir = NA, final.title = NA,
  nocopy = TRUE, nobib = TRUE, doctype = "html", bibtex.dir, bibtex.file,
  csl.file, code.dir, header.file)
}
\description{
Render the Rmarkdown file and move the output
}
