% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send-email.R
\name{send_email}
\alias{send_email}
\title{Send email to alert about some error}
\usage{
send_email(msg, subject)
}
\arguments{
\item{msg}{the message to send}

\item{subject}{subject of the message}
}
\description{
use `gmailr` to send email
}
