% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{appendIfFile}
\alias{appendIfFile}
\title{Write lines to a file if the connection is not null
Else just cat the lines}
\usage{
appendIfFile(lines, file)
}
\description{
Write lines to a file if the connection is not null
Else just cat the lines
}
