% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-packages.R
\name{find_packages}
\alias{find_packages}
\title{Search through files for packages used; make sure DESCRIPTION is up to date.}
\usage{
find_packages(input_dir, output_dir)
}
\description{
Search through files for packages used; make sure DESCRIPTION is up to date.
}
