% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{mround}
\alias{mround}
\title{Round to the nearest anything.
http://r.789695.n4.nabble.com/Is-there-in-R-a-function-equivalent-to-the-mround-as-found-in-most-spreadsheets-td855895.html}
\usage{
mround(x, base)
}
\description{
Round to the nearest anything.
http://r.789695.n4.nabble.com/Is-there-in-R-a-function-equivalent-to-the-mround-as-found-in-most-spreadsheets-td855895.html
}
