% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{pathPrep}
\alias{pathPrep}
\title{Convert a Windows path on the clipboard to a path accepted by R}
\usage{
pathPrep(path = "clipboard")
}
\description{
Convert a Windows path on the clipboard to a path accepted by R
}
